requirements.txt
fastapi
uvicorn[standard]
moviepy
pillow
python-multipart
pydantic
python-dotenv
requests
imageio[ffmpeg]
numpy
sqlmodel
httpx
pytest
aiohttp
backoff
PyJWT
bcrypt
backoff
python-jose[cryptography]
aiohttp
PyJWT
streamlit
bcrypt